% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acv.R
\name{acv}
\alias{acv}
\title{Anisotropic coefficient of variation (ACV)}
\usage{
acv(GRIDobj)
}
\arguments{
\item{GRIDobj}{\code{GRIDobj}, DEM to be processed}
}
\value{
\code{GRIDobj} with anisotropic coefficient of variation values
}
\description{
The anisotropic coefficient of variation describes the general geometry 
of the local land surface and can be used to distinguish elongated from 
oval landforms.
}
\examples{

## load example data set
data(srtm_bigtujunga30m_utm11)
DEM <- terra::unwrap(srtm_bigtujunga30m_utm11)

## calculate ACV
ACV <- acv(GRIDobj = DEM)

## plot result
plot_GRIDobj(GRIDobj = ACV, col = "viridis")

}
