% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.GRIDobj}
\alias{aggregate.GRIDobj}
\title{Aggregate a GRIDobj or point data set to the resolution of a GRIDobj}
\usage{
\method{aggregate}{GRIDobj}(grid1, grid2, xy, xyz, aggfun = "mean")
}
\arguments{
\item{grid1}{\code{GRIDobj} that will be aggregated}

\item{grid2}{\code{GRIDobj} of coarser resolution to be aggregated to}

\item{xy}{\code{data.frame} or \code{matrix} with x and y coordinates
of irregular spaced geodata points. The output will be converted to a
spatial grid of the same resolution and CRS as B. Grid values will be
assigned \code{TRUE}.}

\item{xyz}{\code{data.frame} or \code{matrix} with x, y and z coordinates
of irregular spaced geodata points. The output will be converted to a
spatial grid of the same resolution and CRS as B.}

\item{aggfun}{\code{Character} value, name of the function that is used
to calculate the new aggregated values. Keywords may be
\code{"mean"}, \code{"median"}, \code{"sd"}, \code{"min"}, \code{"max"}.
Default is \code{"mean"}.}
}
\value{
Returns an aggregated \code{GRIDobj}
}
\description{
This function resamples the grid grid1 to match the extent and resolution of
grid B. grid2 must spatially overlap with A. By default, the function uses the
mean to calculate new grid values, but any other function that takes
returns a scalar (e.g. median, std, ...) can be used, too.
}
\details{
Values to be aggregated can also be supplied as list of coordinates
(and attributes). This is particularly useful if point density for
each pixel is greater than one.
}
\examples{

\dontrun{

## load example data set
data(srtm_bigtujunga30m_utm11)
DEM <- GRIDobj(srtm_bigtujunga30m_utm11)

## create GRIDobj with 100 m resolution and the same extent as DEM
grid2 <- GRIDobj(terra::rast(res = c(1000, 1000), ext = ext(DEM$raster)))

## aggregate and plot DEM based on coarse GRIDobj grid2
grid3 <- aggregate(DEM, grid2)
plot(grid3)

## create 100 random points and aggregate to geometry of GRIDobj grid2
pts <- data.frame(x = runif(100, ext(DEM$raster)[1], ext(DEM$raster)[2]),
                  y = runif(100, ext(DEM$raster)[3], ext(DEM$raster)[4]),
                  z = rnorm(100))
D <- aggregate(grid2 = grid2, xyz = pts)
plot(D)

}

}
\author{
Michael Dietze, Wolfgang Schwanghart
}
