% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/and.R
\name{and}
\alias{and}
\title{Perform logical AND operation with two GRIDobj data sets}
\usage{
and(...)
}
\arguments{
\item{...}{\code{GRIDobj} (or \code{SpatRaster}) data sets to be handled}
}
\value{
\code{GRIDobj} with logical values
}
\description{
The function performs a logical AND operation with all values of the two
input data sets, which must be GRIDobj objects.
}
\note{
Preserves the type of the first object in case inputs have differing
classes.
}
\examples{

## Load example data set
data(srtm_bigtujunga30m_utm11)
DEM <- GRIDobj(srtm_bigtujunga30m_utm11)

## Test which values of both DEMs are TRUE
DEM_and <- and(DEM > 0, DEM - 500 > 0)
plot(DEM_and)

## Perform operation on three data sets
DEM_and <- and(DEM > 0, DEM - 500 > 0, DEM - 1000 > 0)
plot(DEM_and)

}
\author{
Michael Dietze, Wolfgang Schwanghart
}
