% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any.R
\name{any.GRIDobj}
\alias{any.GRIDobj}
\title{Test if GRIDobj contains any TRUE value}
\usage{
\method{any}{GRIDobj}(grid, ..., na.rm = FALSE)
}
\arguments{
\item{grid}{\code{GRIDobj} | \code{SpatRaster}

Object to be tested}

\item{...}{Additional arguments (ignored)}

\item{na.rm}{logical

If TRUE \code{NA} values are removed before the result is computed}
}
\value{
\code{Logcial} value of the test result
}
\description{
The function checks if there is any TRUE value in the grid
values and returns a logical value on the result.
}
\examples{

## load example data set
data(srtm_bigtujunga30m_utm11)
DEM <- GRIDobj(srtm_bigtujunga30m_utm11)

## test if any FALSE values are contained
any(DEM)

}
\author{
Michael Dietze, Wolfgang Schwanghart
}
