% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computecosts.R
\name{computecosts}
\alias{computecosts}
\title{Compute costs}
\usage{
computecosts(flats, dem, filled)
}
\arguments{
\item{flats}{\code{SpatRaster} with flats}

\item{dem}{\code{SpatRaster} with DEM}

\item{filled}{\code{SpatRaster} with filled DEM}
}
\value{
Returns a \code{GRIDobj}, logical data set with computed costs
}
\description{
The function DOES THIS AND THAT
}
\examples{
## load example data set
## TO BE DONE

}
