% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRIDobj.R
\name{dim.GRIDobj}
\alias{dim.GRIDobj}
\title{Get the dimensions}
\usage{
\method{dim}{GRIDobj}(x)
}
\arguments{
\item{x}{GRIDobj}
}
\value{
numeric vector

Dimensions of the GRIDobj
}
\description{
`dim.GRIDobj` retrieves the  dimensions of the GRIDobj.
}
