% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOWobj.R
\name{dim_cr.FLOWobj}
\alias{dim_cr.FLOWobj}
\title{FLOWobj dimensions}
\usage{
\method{dim_cr}{FLOWobj}(x)
}
\arguments{
\item{x}{FLOWobj

FLOWobj for which to copmute the dimensions}
}
\value{
numeric vector

Dimensions of the grid
}
\description{
`dim_cr.FLOWobj` computes the  dimensions of the underlying grid of the
FLOWobj in the correct order for libtopotoolbox.
}
