% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRIDobj.R
\name{dim_cr.GRIDobj}
\alias{dim_cr.GRIDobj}
\title{Get the dimensions}
\usage{
\method{dim_cr}{GRIDobj}(x)
}
\arguments{
\item{x}{GRIDobj}
}
\value{
numeric vector

Dimensions of the GRIDobj
}
\description{
`dim_cr.GRIDobj` retrieves the  dimensions of the GRIDobj in the correct
order for libtopotoolbox.
}
