% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_cr.matrix.R
\name{dim_cr.matrix}
\alias{dim_cr.matrix}
\title{Matrix dimensions}
\usage{
\method{dim_cr}{matrix}(x)
}
\arguments{
\item{x}{matrix}
}
\value{
numeric vetor

Dimensions of the matrix
}
\description{
`dim_cr.matrix` retrieves the dimensions of a matrix in the correct
order for libtopotoolbox.
}
\keyword{internal}
