% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dim_cr}
\alias{dim_cr}
\title{Get the dimensions}
\usage{
dim_cr(x)
}
\arguments{
\item{x}{matrix | array | SpatRaster | GRIDobj | FLOWobj}
}
\value{
numeric vector

Dimensions of the object
}
\description{
`dim_cr` retrieves the dimensions of an object in the correct order for
libtopotoolbox.
}
