% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_cr.SpatRaster.R
\name{dim_cr.SpatRaster}
\alias{dim_cr.SpatRaster}
\title{SpatRaster dimensions}
\usage{
\method{dim_cr}{SpatRaster}(x)
}
\arguments{
\item{x}{SpatRaster}
}
\value{
integer vector

Dimensions of the SpatRaster
}
\description{
`dim_cr.SpatRaster` retrieves the dimensions of a SpatRaster in the
correct order for libtopotoolbox.
}
