% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callC.R
\name{EuclideanDT_inC}
\alias{EuclideanDT_inC}
\title{EuclideanDT_inC}
\usage{
EuclideanDT_inC(binary_image, rows, cols, output)
}
\value{
Matrix
}
\description{
callC calls the c function hello that prints out hello world. This is only a demo
}
