% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excesstopography.R
\name{excesstopography}
\alias{excesstopography}
\title{Excess topography}
\usage{
excesstopography(dem, threshold_slope)
}
\arguments{
\item{dem}{GRIDobj; Digital elevation model}

\item{threshold_slope}{float | GRIDobj; The threshold
slope. Spatially variable slopes can be provided as a GRIDobj.}
}
\value{
GRIDobj; The reconstructed topography. The excess
    topography is the difference between the original DEM and the
    reconstructed one.
}
\description{
Reconstruct surface with threshold-slope surface
}
\details{
The excess topography (Blöthe et al. 2015) is computed by solving
an eikonal equation (Anand et al. 2023) constrained to lie below
the original DEM. Where the slope of the DEM is greater than the
threshold slope, the eikonal solver limits the output topography to
that slope, but where the slope of the DEM is lower that the
threshold slope, the output follows the DEM.

The eikonal equation is solved using the fast sweeping method (Zhao
2004), which iterates over the DEM in alternating directions and
updates the topography according to an upwind discretization of the
gradient. To constrain the solution by the original DEM, the output
topography is initiated with the DEM and only updates lower than
the DEM are accepted.
}
\examples{

data(srtm_bigtujunga30m_utm11)
DEM <- GRIDobj(srtm_bigtujunga30m_utm11)
DEMext <- excesstopography(DEM, tan(20*pi/180))

}
