% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOWobj.R
\name{ezgetnal.FLOWobj}
\alias{ezgetnal.FLOWobj}
\title{FLOWobj node attribute list}
\usage{
\method{ezgetnal}{FLOWobj}(ttobj, k)
}
\arguments{
\item{ttobj}{FLOWobj
The FLOWobj for which to retrieve the node attribute list}

\item{k}{GridObject or matrix or array or scalar

The object from which node values will be extracted. If
`k` is a `GridObject` or an `ndarray` with the same shape
as this `FlowObject`, then a copy is returned. If it is a
scalar, an `ndarray` with the appropriate shape, filled
with `k`, is returned.}
}
\value{
GRIDobj or array
}
\description{
Retrieve a node attribute list
}
