% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ezgetnal}
\alias{ezgetnal}
\title{Generic function for ezgetnal}
\usage{
ezgetnal(ttobj, k)
}
\arguments{
\item{ttobj}{FLOWobj or STREAMobj

The TopoToolbox object for which to extract the node attribute list.}

\item{k}{GRIDobj | matrix | array | float

The object from which node values will be extracted. If `k` is a `GridObject`
 or a `matrix` or an `array` with the same shape as the underlying DEM of
 this `TTobj`, the node values will be extracted from the grid by indexing.
 If `k` is an array with the same shape as the node attribute list,
 `ezgetnal` returns a copy of `k`. If `k` is a scalar value, `ezgetnal`
 returns an array of the right shape filled with `k`.}
}
\description{
`ezgetnal` retrieves a node attribute list for a TopoToolbox object.
}
