% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillsink.R
\name{fillsink}
\alias{fillsink}
\title{wrap_fillsink}
\usage{
fillsink(dem, bc)
}
\arguments{
\item{dem}{Takes a Spatrast as input}

\item{bc}{Array of the same size as the input matrix}
}
\value{
A Spatrast
}
\description{
This will make the fillsinks funktion available to R from the libtotopotoolbox subdirectory
}
