% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillsinks.R
\name{fillsinks}
\alias{fillsinks}
\title{fillsinks}
\usage{
fillsinks(dem, bc = NULL, hybrid = TRUE)
}
\arguments{
\item{dem}{GRIDobj | SpatRaster

Digital elevation model}

\item{bc}{numeric array | numeric matrix, optional

Boundary conditions for sink filling. `bc` should match the shape of the DEM.
Values of 1 indicate pixels that should be fixed to their values in the
original DEM and values of 0 indicate pixels that should be filled.}

\item{hybrid}{logical, optional

Should hybrid reconstruction algorithm be used to fill sinks? Defaults to
True. Hybrid reconstruction is faster but requires additional memory be
allocated for a queue.}
}
\value{
GRIDobj | SpatRaster

Filled digital elevation model
}
\description{
Fill sinks in the digital elevation model (DEM).
}
\examples{

data(srtm_bigtujunga30m_utm11)
DEM <- GRIDobj(srtm_bigtujunga30m_utm11)
DEM <- fillsinks(DEM)

}
