% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_accumulation.R
\name{flow_accumulation}
\alias{flow_accumulation}
\title{Compute accumulation of flow}
\usage{
flow_accumulation(fd, weights = 1)
}
\arguments{
\item{fd}{FLOWobj

The FLOWobj that will be the basis of the computation.}

\item{weights}{numeric array or matrix, optional

An array of the same shape as the flow grid representing weights for each
cell, or a constant float value used as the weight for all cells.
If `weights=1.0` (default), the flow accumulation is unweighted.
If an ndarray is provided, it must match the shape of the flow grid., by
default 1.0}
}
\value{
GRIDobj

A new GRIDobj containing the flow accumulation grid.
}
\description{
Computes the flow accumulation for a given flow network using optional
weights. The flow accumulation represents the amount of flow each cell
receives from its upstream neighbors.
}
\examples{
\dontrun{
DEM <- terra::rast(system.file("ex/elev.tif",package="terra"))
DEM <- terra::project(DEM,"EPSG:32632",res=90.0)
FD <- FLOWobj(DEM)
FA <- flow_accumulation(FD)
}

}
