% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_routing_d8_edgelist.R
\name{flow_routing_d8_edgelist}
\alias{flow_routing_d8_edgelist}
\title{wrap_flow_routing_d8_edgelist}
\usage{
flow_routing_d8_edgelist(dem, bc = NULL, hybrid = TRUE)
}
\arguments{
\item{dem}{GRIDobj

Digital elevation model}

\item{bc}{numeric array or matrix, optional

Boundary conditions for sink filling. `bc` should match the shape of the DEM.
Values of 1 indicate pixels that should be fixed to their values in the
original DEM and values of 0 indicate pixels that should be filled.}

\item{hybrid}{logical, optional

Should hybrid reconstruction algorithm be used to fill sinks? Defaults to
True. Hybrid reconstruction is faster but requires additional memory be
allocated for a queue.}
}
\value{
A list containing vectors of source and target pixels for each edge.
}
\description{
Compute downstream pixel indices from flow directions.
}
