% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOWobj.R
\name{FLOWobj}
\alias{FLOWobj}
\title{Create instance of a FLOWobj}
\usage{
FLOWobj(dem, bc = NULL, hybrid = TRUE)
}
\arguments{
\item{dem}{GRIDobj

The GRIDobj that will be the basis of the computation.}

\item{bc}{numeric array or matrix, optional

Boundary conditions for sink filling. `bc` must match the size of the DEM.
Values of 1 indicate pixels that should be fixed to their values in the
original DEM and values of 0 indicate pixels that should be filled.}

\item{hybrid}{logical, optional

Should hybrid reconstruction algorithm be used to fill sinks? Defaults to
True. Hybrid reconstruction is faster but requires additional memory be
allocated for a queue.}
}
\value{
FLOWobj

An object containing vectors of source and target pixels for each edge and
an empty SpatRaster storing the spatial information of the original DEM.
}
\description{
`FLOWobj`is the constructor for the FLOWobj. It takes a GRIDobj from
topotoolboxr or a SpatRaster from terra as input, computes flow directions
and returns them as a FLOWobj. A FLOWobj stores flow directions as a
topologically sorted edge list.
}
\details{
Large intermediate arrays are created during the initialization process,
which could lead to issues when using very large DEMs.
}
\examples{

data(srtm_bigtujunga30m_utm11)
DEM <- GRIDobj(srtm_bigtujunga30m_utm11)
FD <- FLOWobj(DEM)

}
