% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRIDobj.R
\name{get_cellsize.GRIDobj}
\alias{get_cellsize.GRIDobj}
\title{Get the cellsize}
\usage{
\method{get_cellsize}{GRIDobj}(x)
}
\arguments{
\item{x}{GRIDobj}
}
\value{
numeric scalar

Cellsize of the GRIDobj
}
\description{
`get_cellsize.GRIDobj` retrieves the cellsize of the GRIDobj.
}
