% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_cellsize}
\alias{get_cellsize}
\title{Get the cellsize}
\usage{
get_cellsize(x)
}
\arguments{
\item{x}{GRIDobj or FLOWobj}
}
\value{
numeric scalar

Cellsize of the TopoToolbox object
}
\description{
`get_cellsize` retrieves the cellsize of a TopoToolbox object.
}
