% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_dims.array}
\alias{get_dims.array}
\title{Array dimensions}
\usage{
\method{get_dims}{array}(x)
}
\arguments{
\item{x}{array

Array for which to compute dimensions}
}
\value{
numeric vector

Dimensions of the array
}
\description{
`get_dims.array` computes the dimensions of an array in the correct order for
libtopotoolbox.
}
