% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOWobj.R
\name{get_dims.FLOWobj}
\alias{get_dims.FLOWobj}
\title{FLOWobj dimensions}
\usage{
\method{get_dims}{FLOWobj}(x)
}
\arguments{
\item{x}{FLOWobj

FLOWobj for which to copmute the dimensions}
}
\value{
numeric vector

Dimensions of the grid
}
\description{
`get_dims.FLOWobj` computes the  dimensions of the underlying grid of the
FLOWobj in the correct order for libtopotoolbox.
}
