% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRIDobj.R
\name{get_dims.GRIDobj}
\alias{get_dims.GRIDobj}
\title{Get the dimensions}
\usage{
\method{get_dims}{GRIDobj}(x)
}
\arguments{
\item{x}{GRIDobj}
}
\value{
numeric vector

Dimensions of the GRIDobj
}
\description{
`get_dims.GRIDobj` retrieves the  dimensions of the GRIDobj in the correct
order for libtopotoolbox.
}
