% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_dims.matrix}
\alias{get_dims.matrix}
\title{Matrix dimensions}
\usage{
\method{get_dims}{matrix}(x)
}
\arguments{
\item{x}{matrix

Matrix for which to compute dimensions}
}
\value{
numeric vector

Dimensions of the matrix
}
\description{
`get_dims.matrix` computes the dimensions of a matrix in the correct order
for libtopotoolbox.
}
