% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_dims}
\alias{get_dims}
\title{Get the dimensions}
\usage{
get_dims(x)
}
\arguments{
\item{x}{matrix | array | SpatRaster | GRIDobj | FLOWobj}
}
\value{
numeric vector

Dimensions of the object
}
\description{
`get_dims` retrieves the dimensions of an object in the correct order for
libtopotoolbox.
}
