% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_dims.SpatRaster}
\alias{get_dims.SpatRaster}
\title{SpatRaster dimensions}
\usage{
\method{get_dims}{SpatRaster}(x)
}
\arguments{
\item{x}{SpatRaster}
}
\value{
integer vector

Dimensions of the SpatRaster
}
\description{
`get_dims.SpatRaster` retrieves the dimensions of a SpatRaster in the correct
order for libtopotoolbox.
}
\note{
This function serves as a bridge until the GRIDobj is implemented.
}
