% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRIDobj.R
\name{get_extent.GRIDobj}
\alias{get_extent.GRIDobj}
\title{Get the extent of a GRIDobj}
\usage{
\method{get_extent}{GRIDobj}(x)
}
\arguments{
\item{x}{GRIDobj}
}
\value{
named numeric vector

xmin, xmax, ymin and ymax of the GRIDobj
}
\description{
`get_extent.GRIDobj` retrieves the extent of the GRIDobj.
}
