% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_extent}
\alias{get_extent}
\title{Get the extent}
\usage{
get_extent(x)
}
\arguments{
\item{x}{GRIDobj or FLOWobj}
}
\value{
numeric vector

Extent of the TopoToolbox object
}
\description{
`get_extent` retrieves the extent of a TopoToolbox object.
}
