% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid_data.R
\name{get_grid_data}
\alias{get_grid_data}
\title{Takes a Spatrast and gives back a list with the values of the Spatrast, its cellsize and dimensions}
\usage{
get_grid_data(r)
}
\arguments{
\item{r}{Takes a Spatrast}
}
\description{
Takes a Spatrast and gives back a list with the values of the Spatrast, its cellsize and dimensions
}
