% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tif_values.R
\name{get_tiff_values}
\alias{get_tiff_values}
\title{Takes a Spatrast and gives back a matrix with the values of the Spatrast}
\usage{
get_tiff_values(tiff)
}
\arguments{
\item{tiff}{Takes a Spatrast}
}
\description{
Takes a Spatrast and gives back a matrix with the values of the Spatrast
}
