% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRIDobj.R
\name{get_values.GRIDobj}
\alias{get_values.GRIDobj}
\title{Get the cell values}
\usage{
\method{get_values}{GRIDobj}(x)
}
\arguments{
\item{x}{GRIDobj}
}
\value{
numeric vector
}
\description{
`get_values.GRIDobj` retrieves the cell values of a GRIDobj in row-major order.
}
