% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_values}
\alias{get_values}
\title{Get cell values}
\usage{
get_values(x)
}
\arguments{
\item{x}{TopoToolbox object}
}
\value{
numeric vector | logical vector

Cell values of the TopoToolbox object
}
\description{
`get_values` retrieves the cell values of a TopoToolbox object.
}
