% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient8.R
\name{gradient8}
\alias{gradient8}
\title{gradient8}
\usage{
gradient8(dem, unit = "tangent", use_mp = 0)
}
\arguments{
\item{dem}{GRIDobj | SpatRaster

Digital elevation model}

\item{unit}{Unit of returned gradient values. Options are:

  - 'tangent': Calculate the gradient as a tangent (default).

  - 'radian': Calculate the gradient in radians.

  - 'degree': Calculate the gradient in degrees.

  - 'sine': Calculate the gradient as the sine of the angle.

  - 'percent': Calculate the gradient as a percentage.}

\item{use_mp}{Logical

If TRUE, use parallel processing for computation (future feature).
Currently not implemented.}
}
\value{
8-connected neighborhood gradient of a digital elevation model
}
\description{
Computes the gradient of a digital elevation model (DEM) using an 8-direction
algorithm. The algorithm returns the maximum downward gradient and has 
positive values only.
}
\examples{

data(srtm_bigtujunga30m_utm11)
DEM <- GRIDobj(srtm_bigtujunga30m_utm11)
G <- gradient8(DEM)
plot(G, range = c(0, 1))

}
