% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRIDobj2ascii.R
\name{GRIDobj2ascii}
\alias{GRIDobj2ascii}
\title{Save GRIDobj as ESRI ArcGIS ASCII file}
\usage{
GRIDobj2ascii(grid, file)
}
\arguments{
\item{grid}{\code{GRIDobj} to be exported}

\item{file}{\code{Character} value, file name and path where to save
the exported \code{GRIDobj}.}
}
\value{
Writes an ESRI ArcGIS ASCII data set as file.
}
\description{
GRIDobj2ascii writes a GRIDobj to an ESRI ArcGIS ASCII file. The
\code{file} must be a character indicating the relative or absolute
file path. The extension should be either \code{.txt} or \code{.asc}.
A potentially existing file will be overwritten without notice.
}
\details{
This function is a convenience function, mainly for mapping the Matlab
Topotoolbox capabilities to R. It wraps the workhorse function
\code{terra::writeRaster()} (see examples) that in turn uses the
[GDAL driver names](https://gdal.org/en/latest/drivers/raster/index.html).
}
\examples{

\dontrun{

## load example data set
data(srtm_bigtujunga30m_utm11)
DEM <- GRIDobj(srtm_bigtujunga30m_utm11)

## save example data set as ASCII file
GRIDobj2ascii(grid = DEM, file = "export/srtm_bigtujunga30m_utm11.txt")

## the same but using the terra function
terra::writeRaster(grid = DEM,
                   filename = "export/srtm_bigtujunga30m_utm11.txt",
                   filetype = "AAIGrid")

}

}
