% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRIDobj2geotiff.R
\name{GRIDobj2geotiff}
\alias{GRIDobj2geotiff}
\title{Save GRIDobj as geotiff file}
\usage{
GRIDobj2geotiff(grid, file)
}
\arguments{
\item{grid}{\code{GRIDobj} to be exported}

\item{file}{\code{Character} value, file name and path where to save
the exported \code{GRIDobj}.}
}
\value{
Writes a geotiff data set as file.
}
\description{
GRIDobj2geotiff writes a GRIDobj to a geotiff file. The
\code{file} must be a character indicating the relative or absolute
file path. The extension should be either \code{.tif} or \code{.tiff}.
A potentially existing file will be overwritten without notice.
}
\details{
This function is a convenience function, mainly for mapping the Matlab
Topotoolbox capabilities to R. It wraps the workhorse function
\code{terra::writeRaster()} (see examples) that in turn uses the
[GDAL driver names](https://gdal.org/en/latest/drivers/raster/index.html).
}
\examples{

\dontrun{
## load example data set
data(srtm_bigtujunga30m_utm11)
DEM <- GRIDobj(srtm_bigtujunga30m_utm11)

## save example data set as ASCII file
GRIDobj2geotiff(grid = DEM, file = "export/srtm_bigtujunga30m_utm11.tif")

## the same but using the terra function
terra::writeRaster(grid = DEM,
                   filename = "export/srtm_bigtujunga30m_utm11.txt",
                   filetype = "GTiff")
}

}
