% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRIDobj2mat.R
\name{GRIDobj2mat}
\alias{GRIDobj2mat}
\title{Convert GRIDobj to a matrix and coordinate vectors}
\usage{
GRIDobj2mat(grid)
}
\arguments{
\item{grid}{\code{GRIDobj} to be converted}
}
\value{
A \code{list} with the elements \code{$X} (vector of X-values),
\code{$Y} (vector of Y-values), \code{$Z} (numeric matrix of grid values).
}
\description{
The function extracts the grid values and coordinates from a \code{GRIDobj}
and returns a list of X and Y coordinates as well as the corresponding grid
values as matrix.
}
\examples{

## load example data set
data(srtm_bigtujunga30m_utm11)
DEM <- GRIDobj(srtm_bigtujunga30m_utm11)

## extract XYZ information
XYZ <- GRIDobj2mat(grid = DEM)

## plot image of DEM
image(x = XYZ$X, y = XYZ$Y, z = t(XYZ$Z), col = terrain.colors(100))

}
\author{
Michael Dietze, Wolfgang Schwanghart
}
