% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRIDobj2pm.R
\name{GRIDobj2pm}
\alias{GRIDobj2pm}
\title{Convert several GRIDobj objects to a predictor matrix}
\usage{
GRIDobj2pm(...)
}
\arguments{
\item{\dots}{\code{GRIDobj} objects to be processed}
}
\value{
A \code{list} with \code{X}, the predictor matrix (i.e.
column-wise organised grid values pruned from NA-values), \code{ix}, an
index vector indicating the positions of the NA-free values in the original
data sets, and \code{txt}, the names of the input data sets.
}
\description{
The function organises all input \code{GRIDobj} values columnwise into
a matrix. It then identifies row (i.e. shared pixels) that exhibit NA
values and removed those. Note that all input data set must match in
terms of spatial extent and resolution.
}
\examples{

\dontrun{
## load example data set
data(srtm_bigtujunga30m_utm11)
DEM <- GRIDobj(srtm_bigtujunga30m_utm11)

## get gradient, curvature and logs of flow accumulation values
G <- gradient8(dem = DEM)
C <- curvature(GRIDobj = DEM)
A <- flowacc(GRIDobj = fillsinks(GRIDobj = DEM))
terra::values(A)[,1] <- log(terra::values(A)[,1])

## build predictor matrix from gradient, curvature and log flow accumulation 
P <- GRIDobj2pm(gradient = G, curvature = C, logflow = A)

## normalise values
P$X <- (P$X - mean(P$X)) / sd(P$X)

## do simple cluster analysis
km <- kmeans(x = P$X, centers = 5)

## assign cluster values to new landform GRIDobj
landforms <- DEM
terra::values(landforms)[P$ix,1] <- km$cluster

## plot classified data set
plot_GRIDobj(landforms)
}

}
\author{
Michael Dietze, Wolfgang Schwanghart
}
