% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRIDobj2rgb.R
\name{GRIDobj2rgb}
\alias{GRIDobj2rgb}
\title{Convert GRIDobj object to RGB matrix}
\usage{
GRIDobj2rgb(GRIDobj, hex = FALSE, col = "terrain.colors", n = 256)
}
\arguments{
\item{GRIDobj}{\code{SpatRaster} to be converted}

\item{hex}{\code{Logical} value, option to return one matrix with 
hexadecimal colour codes instead of three matrices with 8 Bit RGB colour 
values. Default is \code{FALSE}.}

\item{col}{\code{Character} value, name of the colour space to use. 
Default is \code{"terrain.colors"}.}

\item{n}{\code{Numeric} value, number of different colours to generate.
Default is \code{256} (8 Bit).}
}
\value{
A \code{character} matrix with hexadecimal RGB values of the 
input data set.
}
\description{
The function converts the values of a \code{GRIDobj} into 8 Bit numeric 
RGB values or hexadexcimal definitions of the colour codes. It requires 
the definition of a colour space (see examples).
}
\details{
Available colourspace keywords are the R palettes (\code{"heat.colors"},
\code{"topo.colors"}, \code{"terrain.colors"}). In addition, keywords for 
the [HCL palettes](https://colorspace.r-forge.r-project.org/articles/hcl_palettes.html) 
can be defined, such as \code{"blues3"}, \code{"BuGn"}, \code{"Viridis"},
\code{"Plasma"}, \code{"Batlow"}.
}
\examples{

\dontrun{

## load example data
data(srtm_bigtujunga30m_utm11)
DEM <- terra::unwrap(srtm_bigtujunga30m_utm11)

## convert DEM to RGB colour information
rgb <- GRIDobj2rgb(GRIDobj = DEM)

## convert DEM to hex code using the heat-colors palette and ten colours
hex <- GRIDobj2rgb(GRIDobj = DEM, hex = TRUE, col = "inferno", n = 10)

## plot image (required package 'grid')
grid::grid.raster(image = t(hex$Z))

}

}
\author{
Michael Dietze, Wolfgang Schwanghart
}
