% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRIDobj2xyz.R
\name{GRIDobj2xyz}
\alias{GRIDobj2xyz}
\title{Convert GRIDobj to a data frame of X, Y and z values}
\usage{
GRIDobj2xyz(grid)
}
\arguments{
\item{grid}{\code{GRIDobj} or \code{SpatRaster} to be converted}
}
\value{
A \code{list} with the elements \code{$X} (vector of X-values),
\code{$Y} (vector of Y-values), \code{$z} (numeric matrix of grid values).
}
\description{
The function extracts the grid values and coordinates from a \code{GRIDobj}
and returns a data frame of tupels.
}
\examples{

## load example data set
data(srtm_bigtujunga30m_utm11)
DEM <- GRIDobj(srtm_bigtujunga30m_utm11)

## extract XYZ information
XYZ <- GRIDobj2xyz(grid = DEM)

## plot elevation profile of the first horizontal line of the DEM
plot(x = XYZ$X[1:1197], y = XYZ$Z[1:1197])

}
\author{
Michael Dietze, Wolfgang Schwanghart
}
