% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwdt.R
\name{gwdt}
\alias{gwdt}
\title{wrap_gwdt}
\usage{
gwdt(dem)
}
\arguments{
\item{dem}{GRIDobj | SpatRaster

Digital elevation model}
}
\value{
GRIDobj | SpatRaster

GWDT distances for each grid cell.
}
\description{
Perform the grey-weighted distance transform (GWDT) on the DEM.
}
