% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwdt_computecosts.R
\name{gwdt_computecosts}
\alias{gwdt_computecosts}
\title{wrap_gwdt_computecosts}
\usage{
gwdt_computecosts(flats, original_dem, filled_dem)
}
\arguments{
\item{flats}{GRIDobj | SpatRaster

Flat pixels as returned by identifyflats()}

\item{original_dem}{GRIDobj | SpatRaster

Raw digital elevation model}

\item{filled_dem}{GRIDobj | SpatRaster

Processed digital elevation model}
}
\value{
GRIDobj | SpatRaster

Costs corresponding to each grid cell in the DEM
}
\description{
Compute the cost array used in the gradient-weighted distance
transform (GWDT) algorithm.
}
