% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_topotoolbox.R
\name{has_topotoolbox}
\alias{has_topotoolbox}
\title{has_topotoolbox}
\usage{
has_topotoolbox()
}
\description{
It is a dummy function that allows us to check for the inclusion of our Src C-functions.
}
