% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tiff.R
\name{import_tiff}
\alias{import_tiff}
\title{Wrapper funktion to import a tiff (might be bloat and should be deleted)}
\usage{
import_tiff(path)
}
\arguments{
\item{path}{Takes a string with the path to a tiff file}
}
\description{
Wrapper funktion to import a tiff (might be bloat and should be deleted)
}
