% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRIDobj.R
\name{info}
\alias{info}
\title{Print key variables and statistics of the GRIDobj}
\usage{
info(grid, na.rm = FALSE, show_crs = FALSE)
}
\arguments{
\item{grid}{GRIDobj}

\item{na.rm}{logical

If TRUE \code{NA} values are removed before the result is computed}

\item{show_crs}{logical

If TRUE the coordinate reference system is displayed as well.}
}
\description{
`info` retrieves metadata of the GRIDobj and computes key statistics, e.g.
mean, standard deviation, ...
}
