% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad.R
\name{pad}
\alias{pad}
\title{Add or remove a border of pixels around a GRIDobj}
\usage{
pad(GRIDobj, px = 1, val)
}
\arguments{
\item{GRIDobj}{\code{SpatRaster} to be processed}

\item{px}{\code{Numeric} value, the number of pixels to add (positive 
number) or remove (negavtive number). Default is \code{1}.}

\item{val}{\code{Numeric} value or \code{Characer} keyword. If a numeric 
value is used, that value is repeated for all marginal pixels. If on out of 
the following keywords is used, the respective function is used to calculate 
the values for the marginal pixels: \code{"mean"} or \code{"median"}.}
}
\value{
A \code{GRIDobj} with a padded margin and updated extent information
}
\description{
This function adds or removes a border of constant values or values 
calculated from a focal filter along all edges of a \code{GRIDobj}. The 
number of iterative steps of adding marginal pixels can be controlled, too. 
If the number is negative, the \code{GRIDobj} is cropped by the number of 
defined pixels at each grid border.
}
\examples{

## load example data set
data(srtm_bigtujunga30m_utm11)
DEM <- terra::unwrap(srtm_bigtujunga30m_utm11)

## pad with zeros
DEM_0 <- pad(GRIDobj = DEM)

## pad with mean values of surrounding pixels
DEM_mn <- pad(GRIDobj = DEM, val = "mean")

## pad with two boundaries of NA values
DEM_2na <- pad(GRIDobj = DEM, px = 2, val = NA)

## remove ten sets of pixels
DEM_10off <- pad(GRIDobj = DEM, px = -10)

}
\author{
Michael Dietze, Wolfgang Schwanghart
}
