% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GRIDobj.R
\name{plot_GRIDobj}
\alias{plot_GRIDobj}
\title{Plot a GRIDobj}
\usage{
plot_GRIDobj(GRIDobj, interactive = FALSE, ...)
}
\arguments{
\item{GRIDobj}{\code{GRIDobj} object to plot}

\item{interactive}{\code{Logical} value, option to create an interactive 
plot via \code{plotly} instead of a static map. Default is \code{FALSE}. 
The optiona is only useful if the dataset is a DEM, hence the z-values
are elevation.}

\item{\dots}{Further arguments passed to the plot function}
}
\value{
Graphic output of a spectrogram.
}
\description{
The function plots a GRIDobj. The data structure is converted to a terra 
SpatRaster object and then passed to the standard terra plot routine.
}
\details{
For the interactive case, the following additional arguments may be useful:

1) \code{agg}, aggregation factor to reduce the resolution of the GRIDobj 
   to plot, e.g. use \code{agg = 10} to reduce the size by factor 10
   
2) \code{exa}, exaggeration factor for the elevation, which is useful to 
   improve visibility of topography. Default is \code{1} (no exaggeration).
}
\examples{

## load example data set
data(srtm_bigtujunga30m_utm11)
srtm_bigtujunga30m_utm11 <- terra::unwrap(srtm_bigtujunga30m_utm11)

## plot data set as2D map
plot_GRIDobj(GRIDobj = srtm_bigtujunga30m_utm11)

## plot data set as 3D interactive scene (requires package plotly), note 
## that the scence is aggregated by factor 10 and exaggerated by factor 2
# plot_GRIDobj(GRIDobj = srtm_bigtujunga30m_utm11, 
#              interactive = TRUE, agg = 10, exa = 2)

## plot data set in customised colour
plot_GRIDobj(GRIDobj = srtm_bigtujunga30m_utm11, col = topo.colors(200))

## plot data set without legend
plot_GRIDobj(GRIDobj = srtm_bigtujunga30m_utm11, legend = FALSE)


}
