% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{processgrid}
\alias{processgrid}
\title{Checks a SpatRaster for GRIDobj requirements}
\usage{
processgrid(grid)
}
\arguments{
\item{grid}{GRIDobj | SpatRaster

Input must be either a GRIDobj or SpatRaster}
}
\value{
list

A list with the components 'r', the SpatRaster corresponding to the input
and 'provided_go', a logical value stating whether the input was a GRIDobj
}
\description{
`validateraster` checks whether a provided raster fulfills the requirements
of a TopoToolbox GRIDobj.
}
