% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{randomize}
\alias{randomize}
\title{randomize}
\usage{
randomize(x, ...)
}
\arguments{
\item{x}{\code{GRIDobj}}

\item{\dots}{Further arguments passed to the runif function}
}
\value{
GRIDobj
}
\description{
The function returns a GRIDobj with random uniform number.
}
\examples{

\dontrun{

## load example data set
data(srtm_bigtujunga30m_utm11)
DEM <- GRIDobj(srtm_bigtujunga30m_utm11)
R <- randomize(DEM)
}

}
\author{
Wolfgang Schwanghart
}
