% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRIDobj.R
\name{reproject}
\alias{reproject}
\title{Change the projection}
\usage{
reproject(grid, target, ...)
}
\arguments{
\item{grid}{GRIDobj}

\item{target}{character | GRIDobj | SpatRaster}

\item{...}{Further arguments

See \link[terra]{project}}
}
\value{
GRIDobj `x` with projection from `y`
}
\description{
`reproject` changes the projection of a GRIDobj.
}
