% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOWobj.R
\name{source_indices.FLOWobj}
\alias{source_indices.FLOWobj}
\title{Unravel source indices}
\usage{
\method{source_indices}{FLOWobj}(ttobj)
}
\arguments{
\item{ttobj}{FLOWobj

FLOWobj for which to compute the grid indices of the source pixels}
}
\value{
matrix

Row indices and column indices of the source pixels
}
\description{
`source_indices.FLOWobj` computes row and column indices of the sources of
each edge in the flow network.
}
