% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{source_indices}
\alias{source_indices}
\title{Generic function for source_indices}
\usage{
source_indices(ttobj)
}
\arguments{
\item{ttobj}{FLOWobj or STREAMobj
TopoToolbox object containing the one-dimensional source indices}
}
\value{
n x 2 matrix

Row and column indices of each source pixel in the grid.
}
\description{
`source_indices` uses `unravel_index` to compute the grid indices of
source pixels of a TopoToolbox object.
}
\note{
`source_indices` currently only supports the FLOWobj.
}
