% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{srtm_bigtujunga30m_utm11}
\alias{srtm_bigtujunga30m_utm11}
\title{30 m SRTM DEM for Big Tujunga catchment}
\format{
srtm_bigtujunga30m_utm11
Packed SpatRaster list created by terra::wrap().
}
\source{
USGS SRTM
}
\usage{
srtm_bigtujunga30m_utm11
}
\description{
Wrapped SpatRaster of 30 m SRTM elevation data (643 × 1197 cells, 1 layer,
UTM 11N EPSG:32611) for Big Tujunga, California.
}
\examples{
library(topotoolboxr)
data(srtm_bigtujunga30m_utm11)
dem <- GRIDobj(srtm_bigtujunga30m_utm11)
plot(dem)

}
\keyword{datasets}
