% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOWobj.R
\name{target_indices.FLOWobj}
\alias{target_indices.FLOWobj}
\title{Unravel target indices}
\usage{
\method{target_indices}{FLOWobj}(ttobj)
}
\arguments{
\item{ttobj}{FLOWobj

FLOWobj for which to compute the grid indices of the target pixels}
}
\value{
n x 2 matrix

Row indices and column indices of the target pixels
}
\description{
``target_indices.FLOWobj` computes row and column indices of the targets of
each edge in the flow network.
}
