% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{target_indices}
\alias{target_indices}
\title{Generic function for target_indices}
\usage{
target_indices(ttobj)
}
\arguments{
\item{ttobj}{FLOWobj or STREAMobj
TopoToolbox object containing the one-dimensional target indices}
}
\value{
n x 2 matrix

Row and column indices of each target pixel in the grid
}
\description{
`target_indices` uses `unravel_index` to compute the grid indices of target
pixels of a TopoToolbox object.
}
\note{
`target_indices` currently only supports the FLOWobj.
}
