% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf.R
\name{tf}
\alias{tf}
\title{Test if GRIDobj contains no FALSE (or zero) value}
\usage{
tf(GRIDobj)
}
\arguments{
\item{GRIDobj}{\code{SpatRaster} to be tested}
}
\value{
\code{Logcial} value of the test result
}
\description{
The function checks if there is any FALSE or zero value in the grid 
values and returns a logical value on the result.
}
\examples{

## load example data set
data(srtm_bigtujunga30m_utm11)
DEM <- terra::unwrap(srtm_bigtujunga30m_utm11)

## test if any FALSE values are contained
tf(GRIDobj = DEM)

## introduce FALSE values and test again
terra::values(DEM)[10,] <- FALSE
tf(GRIDobj = DEM)

}
\author{
Michael Dietze, Wolfgang Schwanghart
}
