% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOWobj.R
\name{unravel_index.FLOWobj}
\alias{unravel_index.FLOWobj}
\title{Unravel indices}
\usage{
\method{unravel_index}{FLOWobj}(ttobj, idxs)
}
\arguments{
\item{ttobj}{FLOWobj

FLOWobj for which to compute the grid indices}

\item{idxs}{vector or matrix or array

Flat indices to convert to grid indices}
}
\value{
n x 2 matrix

Row indices and column indices of the provided flat indices
}
\description{
`unravel_index.FLOWobj` unravels the provided linear indices so they can be
used to index grids.
}
