% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{unravel_index}
\alias{unravel_index}
\title{Generic function for unravel_index}
\usage{
unravel_index(ttobj, idxs)
}
\arguments{
\item{ttobj}{FLOWobj or STREAMobj

Object from which the array dimensions are computed}

\item{idxs}{vector or matrix or array

One-dimensional indices to convert into grid indices}
}
\value{
n x 2 matrix

Row and column indices of each pixel in the corresponding array
}
\description{
`unravel_index` converts flat indices into a grid grid indices.
}
\note{
`unravel_index` currently only supports the FLOWobj.
}
