% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validatealignment.R
\name{validatealignment}
\alias{validatealignment}
\title{Check whether multiple objects are aligned}
\usage{
validatealignment(...)
}
\arguments{
\item{...}{matrix | array | GRIDobj | FLOWobj | STREAMobj

Objects to test for alignment}
}
\value{
logical

\code{TRUE} if the two objects are aligned, \code{FALSE} otherwise
}
\description{
`validate_alignment` checks that the objects have the same
`shape` attribute and, if coordinate information is available, the
same coordinate system.
}
\examples{
\dontrun{
M <- matrix(1:25, 5, 5)
DEM <- terra::rast(M, crs="EPSG:25833")
FD <- FLOWobj(DEM)
print(validatealignment(M, DEM, FD))
}

}
