% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateraster.R
\name{validateraster}
\alias{validateraster}
\title{Checks a SpatRaster for GRIDobj requirements}
\usage{
validateraster(raster)
}
\arguments{
\item{raster}{character | SpatRaster

The input can be either a path to a file containing the raster or a
SpatRaster from terra}
}
\value{
logical

Returns TRUE if the raster fulfills the requirements
}
\description{
`validateraster` checks whether a provided raster fulfills the requirements
of a TopoToolbox GRIDobj.
}
