% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_fillsink.R
\name{wrap_fillsink}
\alias{wrap_fillsink}
\title{wrap_fillsink}
\usage{
wrap_fillsink(tiff, bc)
}
\arguments{
\item{tiff}{Takes a Spatrast as input}
}
\value{
An array
}
\description{
This will make the fillsinks funktion available to R from the libtotopotoolbox subdirectory
}
