% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_gradient8.R
\name{wrap_gradient8}
\alias{wrap_gradient8}
\title{wrap_gradient8}
\usage{
wrap_gradient8(output, dem, cellsize, use_mp, dims)
}
\description{
This will make the Gradient8 funktion available to R from the libtotopotoolbox subdirectory
}
