% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillsinks.R
\name{fillsinks}
\alias{fillsinks}
\title{Fill pits, sinks or topographic depressions}
\usage{
fillsinks(GRIDobj, maxdepth, sinks, B)
}
\arguments{
\item{GRIDobj}{\code{SpatRaster} with DEM}

\item{maxdepth}{\code{Numerical} value, maximum depth of sinks that will 
be filled}

\item{sinks}{\code{Logical} matrix or \code{GRIDobj} of the same size as 
the DEM. \code{TRUE} values in sinks are treated as sinks in the digital 
elevation model and are not filled.}

\item{B}{\code{SpatRaster} or boundary matrix.}
}
\value{
Returns a \code{GRIDobj}, digital elevation model with filled sinks
}
\description{
The function removes topographic depressions in a Digital Elevation Model 
(DEM). Use this function to enable a continuous flow towards the DEM edges.
}
\details{
OPEN QUESTIONS:

1) is Matlabs Image Processing Toolbox still an option under R?

2) is implementation of -999 versus -Inf for NA values correct?

3) What is the meaning of B (bc)? Does it replace sinks and maxdepth?

Sinks are typically artefacts emerging by the generation of the DEM.
However, there may also be closed basins or sinkholes and as such those 
sinks are important real features of DEMs. In order to account for such 
sinks, \code{fillsinks()} allows to specify a maximum depth of sinks, that 
will be filled, or to employ a logical data set (\code{sinks}) that is 
\code{TRUE} where sinks should remain (minima imposition). Note that for 
the latter option, there will be one regional minima for each connected 
component in the sinks data set.
}
\examples{
## load example data set
data(srtm_bigtujunga30m_utm11)
DEM <- terra::unwrap(srtm_bigtujunga30m_utm11)
                                         
## fill sinks
DEM_filled <- fillsinks(GRIDobj = DEM)

## plot difference
plot_GRIDobj(DEM_filled - DEM)

}
