% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_routing_d8_carve.R
\name{flow_routing_d8_carve}
\alias{flow_routing_d8_carve}
\title{wrap_flow_routing_d8_carve}
\usage{
flow_routing_d8_carve(DEM, bc = NULL, hybrid = TRUE)
}
\arguments{
\item{DEM}{GRIDobj

Digital elevation model}

\item{bc}{numeric array or matrix, optional

Boundary conditions for sink filling. `bc` should match the shape of the DEM.
Values of 1 indicate pixels that should be fixed to their values in the
original DEM and values of 0 indicate pixels that should be filled.}

\item{hybrid}{logical, optional

Should hybrid reconstruction algorithm be used to fill sinks? Defaults to
True. Hybrid reconstruction is faster but requires additional memory be
allocated for a queue.}
}
\value{
A list containing two terra::SpatRaster representing the source cells
for flow routing (source) and the flow direction for each grid cell (direction).
}
\description{
Compute the flow routing using the D8 algorithm with carving for flat areas.
}
