% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_routing_d8_carve.R
\name{flow_routing_d8_carve}
\alias{flow_routing_d8_carve}
\title{wrap_flow_routing_d8_carve}
\usage{
flow_routing_d8_carve(DEM)
}
\arguments{
\item{DEM}{(terra::SpatRaster) Digital elevation model}
}
\value{
A list containing two terra::SpatRaster representing the source cells
for flow routing (source) and the flow direction for each grid cell (direction).
}
\description{
Compute the flow routing using the D8 algorithm with carving for flat areas.
}
