% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_routing_d8_edgelist.R
\name{flow_routing_d8_edgelist}
\alias{flow_routing_d8_edgelist}
\title{wrap_flow_routing_d8_edgelist}
\usage{
flow_routing_d8_edgelist(DEM)
}
\arguments{
\item{DEM}{(terra::SpatRaster) Digital elevation model}
}
\value{
A list containing vectors of source and target pixels for each edge.
}
\description{
Compute downstream pixel indices from flow directions.
}
