% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_dims}
\alias{get_dims}
\title{Generic function for computing dimensions}
\usage{
get_dims(x)
}
\arguments{
\item{x}{matrix or array or SpatRaster or GRIDobj or FLOWobj

Object for which to compute the dimensions}
}
\value{
numeric vector

Dimensions of the object
}
\description{
`get_dims` computes the dimensions of an object in the correct order for
libtopotoolbox.
}
