% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_grid_data}
\alias{get_grid_data}
\title{Get grid data}
\usage{
get_grid_data(dem)
}
\arguments{
\item{dem}{GRIDobj | SpatRaster

Object for which to retrieve grid data}
}
\value{
list

A list with the components 'z', a vector containing the grid data in
row-major order, 'cellsize', the horizontal resolution of the grid, and
'dims' a two-element vector with the number of columns and rows of the grid.
}
\description{
`get_grid_data` retrieves the values, cellsize and dimension of a provided
GRIDobj or SpatRaster.
}
