% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient8.R
\name{gradient8}
\alias{gradient8}
\title{wrap_gradient8}
\usage{
gradient8(dem, use_mp = 0)
}
\arguments{
\item{dem}{Input of type Spatrast from Terra}

\item{use_mp}{Future feature that will allow to parallelize the code}
}
\value{
A Spatrast
}
\description{
This will make the Gradient8 funktion available to R from the libtotopotoolbox subdirectory
}
