% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwdt.R
\name{gwdt}
\alias{gwdt}
\title{wrap_gwdt}
\usage{
gwdt(DEM)
}
\arguments{
\item{DEM}{(terra::SpatRaster) Digital elevation model}
}
\value{
A terra::SpatRaster representing the GWDT distances for each grid cell.
}
\description{
Perform the grey-weighted distance transform (GWDT) on the DEM.
}
