% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwdt.R
\name{gwdt}
\alias{gwdt}
\title{gwdt}
\usage{
gwdt(dem)
}
\arguments{
\item{dem}{GRIDobj | SpatRaster

Digital elevation model}
}
\value{
GRIDobj | SpatRaster

GWDT distances for each grid cell.
}
\description{
Perform the grey-weighted distance transform (GWDT) on the DEM. The function
is for internal use and calculates the auxiliary topography in topographic
sinks and flats that is required to derive flow directions.
}
\details{
gwdt fills topographic sinks, identifies flat areas and sill pixels, and
then calculates the gray-weighted distance transform on flat areas seeded
at sill pixels.
}
