% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwdt_computecosts.R
\name{gwdt_computecosts}
\alias{gwdt_computecosts}
\title{wrap_gwdt_computecosts}
\usage{
gwdt_computecosts(flats, original_dem, filled_dem)
}
\arguments{
\item{flats}{}

\item{original_dem}{(terra::SpatRaster) Raw digital elevation model}

\item{filled_dem}{(terra::SpatRaster) Processed DEM}
}
\value{
List containing costs and connected components
}
\description{
This function makes the libopotoolbox cost computation for the gray-weighted 
distance transform available in R
}
