% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwdt_computecosts.R
\name{gwdt_computecosts}
\alias{gwdt_computecosts}
\title{wrap_gwdt_computecosts}
\usage{
gwdt_computecosts(flats, original_dem, filled_dem)
}
\arguments{
\item{flats}{(terra::SpatRaster) Flat pixels as returned by identifyflats()}

\item{original_dem}{(terra::SpatRaster) Raw DEM}

\item{filled_dem}{(terra::SpatRaster) Processed DEM}
}
\value{
A terra::SpatRaster of costs corresponding to each grid cell in the DEM.
}
\description{
Compute the cost array used in the gradient-weighted distance
transform (GWDT) algorithm.
}
