% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyflats.R
\name{identifyflats}
\alias{identifyflats}
\title{Identify flats in elevation data}
\usage{
identifyflats(GRIDobj)
}
\arguments{
\item{GRIDobj}{\code{SpatRaster} with DEM}
}
\value{
Returns a \code{GRIDobj}, logical data set with flats
}
\description{
The function DOES THIS AND THAT
}
\examples{
## load example data set
data(srtm_bigtujunga30m_utm11)
DEM <- terra::unwrap(srtm_bigtujunga30m_utm11)
                                         
## TO BE DONE

}
