% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyflats.R
\name{identifyflats}
\alias{identifyflats}
\title{wrap_identifyflats}
\usage{
identifyflats(dem)
}
\arguments{
\item{dem}{Input of type Spatrast from Terra}
}
\value{
A Spatrast
}
\description{
This will make the identifyflats function available to R from the libtotopotoolbox subdirectory
}
