% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyflats.R
\name{identifyflats}
\alias{identifyflats}
\title{identifyflats}
\usage{
identifyflats(dem)
}
\arguments{
\item{dem}{GRIDobj | SpatRaster

Digital elevation model}
}
\value{
GRIDobj | SpatRaster

Grid of identified flats, sills and presills
}
\description{
This function identifies flat pixels in a digital elevation model. 
A flat pixel is one surrounded by pixels with the same or higher 
elevations and has the value 1 in the output grid. In addition, 
the function identifies also sill (value = 2) and presill pixels 
(value = 5). presill pixels are located next to sill pixels and
have the same elevation. The function uses libtopotoolbox.
}
\examples{

data(srtm_bigtujunga30m_utm11)
DEM <- GRIDobj(srtm_bigtujunga30m_utm11)
DEMf <- fillsinks(DEM)
I <- identifyflats(DEMf)
plot(I)

}
