% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRIDobj.R
\name{info}
\alias{info}
\title{Print key variables and statistics of the GRIDobj}
\usage{
info(x)
}
\arguments{
\item{x}{GRIDobj}
}
\description{
`info` retrieves metadata of the GRIDobj and computes key statistics, e.g.
mean, standard deviation, ...
}
