% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRIDobj.R
\name{plot.GRIDobj}
\alias{plot.GRIDobj}
\title{Plot a GRIDobj}
\usage{
\method{plot}{GRIDobj}(x, ...)
}
\arguments{
\item{x}{GRIDobj}

\item{...}{Further arguments

See also \link[graphics]{plot}}
}
\description{
`plot.GRIDobj` plots the values of a GRIDobj.
}
