% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRIDobj.R
\name{reproject}
\alias{reproject}
\title{Change the projection}
\usage{
reproject(x, y, ...)
}
\arguments{
\item{x}{GRIDobj}

\item{y}{character | GRIDobj}

\item{...}{Further arguments}
}
\value{
GRIDobj `x` with projeciton from `y`

See also \link[terra]{project}
}
\description{
`reproject` changes the projection of a GRIDobj.
}
