% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topotoolbox-package.R
\docType{data}
\name{srtm_bigtujunga30m_utm11}
\alias{srtm_bigtujunga30m_utm11}
\title{A digital elevation model of the Big Tujunga catchment}
\format{
A SpatRaster with 643 rows and 1197 columns, referenced in the 
CRS UTM zone 11N.
}
\usage{
data(srtm_bigtujunga30m_utm11)
}
\description{
The dataset contains a wrapped SpatRaster data set, representing a 
digital elevation model (DEM) of the Big Tujunga catchment, USA. After 
loading the example data, it needs to be unwrapped in order to use it.
}
\keyword{datasets}
